refactor: :recycle: move addressSchema to channels-config.ts
