--- apps/taxes/src/modules/taxes/tax-common-schema.ts
@@ -1,11 +1,0 @@
-import { z } from "zod";
-
-export const addressSchema = z.object({
-  country: z.string(),
-  zip: z.string(),
-  state: z.string(),
-  city: z.string(),
-  street: z.string(),
-});
-
-export type ChannelAddress = z.infer<typeof addressSchema>;
